pyproject.toml
[tool.poetry]
name = "portfolio"
version = "0.1.0"
description = ""
authors = ["Aman Rawat <ar.rawataman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
django-tailwind = "^3.8.0"
django-htmx = "^1.17.2"
django-autoslug = "^1.9.9"
requests = "^2.31.0"
dj-database-url = "^2.1.0"
whitenoise = {extras = ["brotli"], version = "^6.6.0"}
gunicorn = "^21.2.0"
pillow = "^10.2.0"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.2.0"
python-dotenv = "^1.0.0"
django-browser-reload = "^1.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
